compute_sam {pack}	R Documentation

Compute the value of sam between case and control

Description

Compute the value of sam between case and control

Usage

compute_sam(m, n, delta)

Arguments

m
A numeric variable

n
A numeric variable

delta
A value that adjusts the false discovery rate(FDR)

Value

m*n sam results

Examples

compute_sam(100,20,0.000000000000000001)

[Package pack version 0.1.0 Index]